refactor: move/rename logic used to export to engines
